requirements_web.txt
# VNU-UET Research Hours Web Application

Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
Flask-Limiter>=3.5.0
Werkzeug>=3.0.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.6
Pillow>=10.0.0

# Optional: for production deployment
gunicorn>=21.0.0
python-dotenv>=1.0.0
